requirements.txt
gTTS
elevenlabs
pydub
python-dotenv
gradio
groq
